requirements.txt
aiosqlite==0.16.1
appdirs==1.4.4
asyncpg==0.22.0
bcrypt==3.2.0
black==20.8b1
cffi==1.14.5
click==7.1.2
Deprecated==1.2.12
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
fastapi-users==5.1.3
h11==0.12.0
httptools==0.1.1
idna==3.1
iso8601==0.1.14
makefun==1.11.2
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.8.1
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
pypika-tortoise==0.1.0
python-dotenv==0.17.0
python-multipart==0.0.5
pytz==2020.5
PyYAML==5.4.1
regex==2021.4.4
six==1.15.0
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.17.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
wrapt==1.12.1
